requirements.txt
anyio==4.9.0
bcrypt==3.2.2
cryptography==45.0.2
fastapi==0.115.12
idna==3.10
motor==3.7.1
passlib==1.7.4
pydantic==2.11.4
pydantic_core==2.33.2
pymongo==4.13.0
python-dotenv==1.1.0
python-jose==3.4.0
python-multipart==0.0.20
setuptools==80.9.0
sniffio==1.3.1
starlette==0.46.2
uvicorn==0.34.2
